requirements.txt
PySide6>=6.8.0
PyMuPDF>=1.25.0
pypdf>=5.0.0
pytesseract>=0.3.13
Pillow>=10.4.0
opencv-python>=4.10.0.84
pytest>=8.3.0
PyInstaller>=6.11.0
